Add changes to column value of table ratings
- Changed type from int to tinyint
- Added a check constraint to limit the accepted values to 1-5

+++ SteamMates/Models/Persistence/Rating.cs
@@ -16,6 +16,6 @@ namespace SteamMates.Models.Persistence
         [ForeignKey(nameof(Game))]
         public long GameId { get; set; }
 
-        public int Value { get; set; }
+        public byte Value { get; set; }
     }
 }
